Move and document StateMachine Helper
